가계부 테스트 페이지 및 가계부 카드 컴포넌트 테스트 삭제 및 코드 수정
