+++ src/components/index.jsx
@@ -2,7 +2,6 @@ export { default as DefaultBtn } from "components/ui/button/DefaultBtn";
 export { default as Nav } from "components/common/Navigation";
 
 export { default as CashBookCard } from "components/cash-book/CashBookCard"
-export { default as CashBookCardTemp } from "components/cash-book/CashBookCardTemp"
 export { default as DayPicker } from "components/cash-book/DayPicker"
 export { default as DayWrapper } from "components/cash-book/DayWrapper"
 export { default as CashBookDetailList } from "components/cash-book/CashBookDetailList";
